refactor(orchestrator): extract persistence-index bookkeeping into orchestration/persistence-index.ts (seam 23, #1526)
Behavior-preserving move of content-hash dedup index ops, bitemporal
backfill-on-dedup-hit, temporal tag index updates, persisted-memory
indexing, graph edge construction, and semantic dedup lookup (433 LOC,
9 methods, 8 deps) into PersistenceIndexCoordinator (late-binding deps
rule, seams 18-22). resolveRecentThreadMemoryPaths exported.

orchestrator.ts: 6,772 -> 6,476 LOC.
